requirements.txt
asgiref==3.11.0
attrs==25.4.0
dj-database-url==3.0.1
Django==5.2.8
django-cors-headers==4.9.0
django-environ==0.12.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.11
gunicorn==22.0.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
psycopg2-binary==2.9.11
PyJWT==2.10.1
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
uritemplate==4.2.0
